requirements.txt



Django==3.0.2
django-admin-bootstrapped==2.5.7
djangorestframework==3.11.0